64-bit network types should be long long int -This doesn't guarantee 8 bytes, but most implementations I want to support are 8 bytes

+++ include/GFK/Network/BytePacker.hpp
@@ -7,11 +7,11 @@ class BytePacker
 {
 public:
 	static unsigned int PackFloat32(float f);
-	static unsigned long int PackFloat64(double f);
+	static unsigned long long int PackFloat64(double f);
 	static float UnpackFloat32(unsigned int packedFloat);
-	static double UnpackFloat64(unsigned long int packedDouble);
-	static unsigned long int htonll(unsigned long int value);
-	static unsigned long int ntohll(unsigned long int value);
+	static double UnpackFloat64(unsigned long long int packedDouble);
+	static unsigned long long int htonll(unsigned long long int value);
+	static unsigned long long int ntohll(unsigned long long int value);
 private:
 	static unsigned long long int Pack754(long double f, unsigned bits, unsigned expbits);
 	static long double Unpack754(unsigned long long int i, unsigned bits, unsigned expbits);
