+++ src/GFK/Network/BytePacker.cpp
@@ -101,7 +101,7 @@ unsigned int BytePacker::PackFloat32(float f)
 	return Pack754(f, 32, 8);
 }
 
-unsigned long int BytePacker::PackFloat64(double d)
+unsigned long long int BytePacker::PackFloat64(double d)
 {
 	return Pack754(d, 64, 11);
 }
@@ -111,13 +111,13 @@ float BytePacker::UnpackFloat32(unsigned int packedFloat)
 	return Unpack754(packedFloat, 32, 8);
 }
 
-double BytePacker::UnpackFloat64(unsigned long int packedDouble)
+double BytePacker::UnpackFloat64(unsigned long long int packedDouble)
 {
 	return Unpack754(packedDouble, 64, 11);
 }
 
 // From http://stackoverflow.com/questions/3022552/is-there-any-standard-htonl-like-function-for-64-bits-integers-in-c
-unsigned long int BytePacker::htonll(unsigned long int value)
+unsigned long long int BytePacker::htonll(unsigned long long int value)
 {
 	static const int num = 42;
 
@@ -127,7 +127,7 @@ unsigned long int BytePacker::htonll(unsigned long int value)
 		const uint32_t high_part = htonl(static_cast<uint32_t>(value >> 32));
 		const uint32_t low_part = htonl(static_cast<uint32_t>(value & 0xFFFFFFFFL));
 
-		return (static_cast<unsigned long int>(low_part) << 32) | high_part;
+		return (static_cast<unsigned long long int>(low_part) << 32) | high_part;
 	}
 	else
 	{
@@ -136,7 +136,7 @@ unsigned long int BytePacker::htonll(unsigned long int value)
 }
 
 // From http://stackoverflow.com/questions/3022552/is-there-any-standard-htonl-like-function-for-64-bits-integers-in-c
-unsigned long int BytePacker::ntohll(unsigned long int value)
+unsigned long long int BytePacker::ntohll(unsigned long long int value)
 {
 	static const int num = 42;
 
@@ -146,7 +146,7 @@ unsigned long int BytePacker::ntohll(unsigned long int value)
 		const uint32_t high_part = htonl(static_cast<uint32_t>(value >> 32));
 		const uint32_t low_part = htonl(static_cast<uint32_t>(value & 0xFFFFFFFFL));
 
-		return (static_cast<unsigned long int>(low_part) << 32) | high_part;
+		return (static_cast<unsigned long long int>(low_part) << 32) | high_part;
 	}
 	else
 	{
